Address review round 4: single BSWAP16 home, mixer arg refactor, stragglers
effects.h includes macros.h instead of carrying its own BSWAP16 copy.
mixer.c drops its unused arguments and locals outright, with the ABI
macros absorbing the removed parameters the same way they already
absorb pkt. func_8028EC98 takes no argument, the fourth intrinsic
pragma leaves math_util_2.c, the backup checksum loop joins its
siblings on size_t, and two leftover casts and an UNUSED go away.

+++ src/save.c
@@ -532,10 +532,10 @@ void update_save_data_backup(void) {
 
 u8 compute_save_data_checksum_backup_1(void) {
     u8* backupGrandPrixPoints = gSaveData.backup.saveInfo.grandPrixPoints;
-    s32 i;
+    size_t i;
     s32 crc = 0;
 
-    for (i = 0; i < (s32) sizeof(SaveInfo); i++) {
+    for (i = 0; i < sizeof(SaveInfo); i++) {
         crc += ((backupGrandPrixPoints[i] + 1) * (i + 1)) + i;
     }
